+++ jdaviz/configs/default/plugins/collapse/tests/test_collapse.py
@@ -22,7 +22,17 @@ def test_linking_after_collapse(cubeviz_helper, spectral_cube_wcs):
 
     assert len(dc) == 2
     assert dc[1].label == 'collapsed'
-    assert len(dc.external_links) == 2
+    assert len(dc.external_links) == 3
 
-    assert dc.external_links[1].cids1[0] is dc[0].pixel_component_ids[1]
-    assert dc.external_links[1].cids2[0] is dc[1].pixel_component_ids[0]
+    # Link 3D z to 2D x and 3D y to 2D y
+
+    # Link 1:
+    # Pixel Axis 0 [z] from cube.pixel_component_ids[0]
+    # Pixel Axis 1 [x] from plugin.pixel_component_ids[1]
+    assert dc.external_links[1].cids1[0] == dc[0].pixel_component_ids[0]
+    assert dc.external_links[1].cids2[0] == dc[-1].pixel_component_ids[1]
+    # Link 2:
+    # Pixel Axis 1 [y] from cube.pixel_component_ids[1]
+    # Pixel Axis 0 [y] from plugin.pixel_component_ids[0]
+    assert dc.external_links[2].cids1[0] == dc[0].pixel_component_ids[1]
+    assert dc.external_links[2].cids2[0] == dc[-1].pixel_component_ids[0]