+++ jdaviz/configs/default/plugins/model_fitting/model_fitting.py
@@ -6,7 +6,6 @@ from specutils import Spectrum1D
 from specutils.utils import QuantityModel
 from traitlets import Bool, List, Unicode, observe
 from glue.core.data import Data
-from glue.core.link_helpers import LinkSame
 
 from jdaviz.core.events import SnackbarMessage
 from jdaviz.core.registries import tray_registry
@@ -577,10 +576,3 @@ class ModelFitting(PluginTemplateMixin, DatasetSelectMixin,
 
         self.add_results.add_results_from_plugin(spectrum)
         self._set_default_results_label()
-
-        # Link the result spectrum to the reference data of the spectrum viewer
-
-        ref_data = self.app.get_viewer('spectrum-viewer').state.reference_data
-        data_id = ref_data.world_component_ids[-1]
-        model_id = self.app.session.data_collection[self.results_label].world_component_ids[0]
-        self.app.session.data_collection.add_link(LinkSame(data_id, model_id))